feat: add external-banking controller and fix decimal types
- Add ExternalBankingController with completeBankingOperation endpoint
- Refactor TinkoffMockService.confirmPayment to accept status parameter
- Change amount/balance fields from decimal/bigint to int (no kopeks)
- Update MockPayment to call new external-banking endpoint
- Fix FanDashboard deposit flow

+++ backend/src/transactions/entities/transaction.entity.ts
@@ -25,12 +25,12 @@ export class Transaction {
   })
   type: TransactionType;
 
-  // Сумма транзакции (в копейках)
-  @Column({ type: 'bigint' })
+  // Сумма транзакции (целое число, без копеек)
+  @Column({ type: 'int' })
   amount: number;
 
-  // Баланс после транзакции
-  @Column({ type: 'bigint' })
+  // Баланс после транзакции (целое число, без копеек)
+  @Column({ type: 'int' })
   balanceAfter: number;
 
   // Описание транзакции
